move fields into ClaimDetailsViewField
