+++ src/main/java/uk/gov/justice/laa/amend/claim/viewmodels/viewfield/ClaimDetailsViewField.java
@@ -3,8 +3,10 @@ package uk.gov.justice.laa.amend.claim.viewmodels.viewfield;
 import java.util.List;
 import java.util.function.Function;
 import lombok.Getter;
+import uk.gov.justice.laa.amend.claim.models.CivilClaimDetails;
 import uk.gov.justice.laa.amend.claim.models.Claim;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
+import uk.gov.justice.laa.amend.claim.models.MediationClaimDetails;
 
 @Getter
 public enum ClaimDetailsViewField implements ClaimViewField<ClaimDetails> {
@@ -20,6 +22,9 @@ public enum ClaimDetailsViewField implements ClaimViewField<ClaimDetails> {
   CASE_REFERENCE_NUMBER(new Accessor<>(Claim::getCaseReferenceNumber)),
   CASE_START_DATE(new Accessor<>(Claim::getCaseStartDate), FieldType.DATE),
   UNIQUE_FILE_NUMBER(new Accessor<>(Claim::getUniqueFileNumber)),
+  FEE_CODE(new Accessor<>(ClaimDetails::getFeeCode)),
+  MATTER_TYPE_CODE_1(new Accessor<>(ClaimDetailsViewField::matterTypeCodeOne)),
+  MATTER_TYPE_CODE_2(new Accessor<>(ClaimDetailsViewField::matterTypeCodeTwo)),
   STAGE_REACHED(new Accessor<>(ClaimDetails::getStageReached), FieldOptions.STAGE_REACHED),
   CASE_CONCLUDED_DATE(new Accessor<>(Claim::getCaseEndDate), FieldType.DATE),
 
@@ -61,4 +66,24 @@ public enum ClaimDetailsViewField implements ClaimViewField<ClaimDetails> {
 
   public record Accessor<T>(Function<ClaimDetails, T> getter)
       implements ClaimViewFieldAccessor<ClaimDetails, T> {}
+
+  private static String matterTypeCodeOne(ClaimDetails claim) {
+    if (claim instanceof CivilClaimDetails civilClaim) {
+      return civilClaim.getMatterType1();
+    }
+    if (claim instanceof MediationClaimDetails mediationClaim) {
+      return mediationClaim.getMatterType1();
+    }
+    return null;
+  }
+
+  private static String matterTypeCodeTwo(ClaimDetails claim) {
+    if (claim instanceof CivilClaimDetails civilClaim) {
+      return civilClaim.getMatterType2();
+    }
+    if (claim instanceof MediationClaimDetails mediationClaim) {
+      return mediationClaim.getMatterType2();
+    }
+    return null;
+  }
 }
